fold eachBatchParallel into eachBatchParallelIngestion
