+++ plugin-server/src/main/ingestion-queues/batch-processing/each-batch.ts
@@ -1,7 +1,6 @@
 import * as Sentry from '@sentry/node'
 import { EachBatchPayload, KafkaMessage } from 'kafkajs'
 
-import { PipelineEvent } from '../../../types' // Must require as `tsc` strips unused `import` statements and just requiring this seems to init some globals
 import { status } from '../../../utils/status'
 import { IngestionConsumer } from '../kafka-queue'
 import { latestOffsetTimestampGauge } from '../metrics'
@@ -73,95 +72,6 @@ export async function eachBatch(
                 new Date().valueOf() - batchStartTimer.valueOf()
             }ms (${loggingKey})`
         )
-    } finally {
-        queue.pluginsServer.statsd?.timing(`kafka_queue.${loggingKey}`, batchStartTimer)
-        transaction.finish()
-    }
-}
-
-export async function eachBatchParallel(
-    /**
-     * Using the provided groupIntoBatches function, split the incoming batch into micro-batches
-     * that are executed **in parallel**, committing offsets only at the end.
-     * Events within a single micro-batch are processed **sequentially**.
-     */
-    { batch, resolveOffset, heartbeat, commitOffsetsIfNecessary, isRunning, isStale }: EachBatchPayload,
-    queue: IngestionConsumer,
-    eachMessage: (message: PipelineEvent, queue: IngestionConsumer) => Promise<void>,
-    groupIntoBatches: (messages: KafkaMessage[]) => PipelineEvent[][],
-    concurrency: number,
-    key: string
-): Promise<void> {
-    const batchStartTimer = new Date()
-    const loggingKey = `each_batch_${key}`
-
-    const transaction = Sentry.startTransaction(
-        { name: `eachBatchParallel(${eachMessage.name})` },
-        { topic: queue.topic }
-    )
-
-    try {
-        /**
-         * Micro-batches should be executed from biggest to smallest to enable the best concurrency.
-         * We're sorting with biggest last and pop()ing. Ideally, we'd use a priority queue by length
-         * and a separate array for single messages, but let's look at profiles before optimizing.
-         */
-        const messageBatches = groupIntoBatches(batch.messages)
-        messageBatches.sort((a, b) => a.length - b.length)
-
-        queue.pluginsServer.statsd?.histogram('ingest_event_batching.input_length', batch.messages.length, { key: key })
-        queue.pluginsServer.statsd?.histogram('ingest_event_batching.batch_count', messageBatches.length, { key: key })
-
-        async function processMicroBatches(batches: PipelineEvent[][]): Promise<void> {
-            let currentBatch
-            let processedBatches = 0
-            while ((currentBatch = batches.pop()) !== undefined) {
-                const batchSpan = transaction.startChild({
-                    op: 'messageBatch',
-                    data: { batchLength: currentBatch.length },
-                })
-                for (const message of currentBatch) {
-                    await Promise.all([eachMessage(message, queue), heartbeat()])
-                }
-                processedBatches++
-                batchSpan.finish()
-            }
-            status.debug('🧩', `Stopping worker after processing ${processedBatches} micro-batches`)
-            return Promise.resolve()
-        }
-
-        await Promise.all([...Array(concurrency)].map(() => processMicroBatches(messageBatches)))
-
-        // Commit offsets once at the end of the batch. We run the risk of duplicates
-        // if the pod is prematurely killed in the middle of a batch, but this allows
-        // us to process events out of order within a batch, for higher throughput.
-        const commitSpan = transaction.startChild({ op: 'offsetCommit' })
-        const lastMessage = batch.messages.at(-1)
-        if (lastMessage) {
-            resolveOffset(lastMessage.offset)
-            await commitOffsetsIfNecessary()
-            latestOffsetTimestampGauge
-                .labels({ partition: batch.partition, topic: batch.topic, groupId: key })
-                .set(Number.parseInt(lastMessage.timestamp))
-        }
-        commitSpan.finish()
-
-        status.debug(
-            '🧩',
-            `Kafka batch of ${batch.messages.length} events completed in ${
-                new Date().valueOf() - batchStartTimer.valueOf()
-            }ms (${loggingKey})`
-        )
-
-        if (!isRunning() || isStale()) {
-            status.info('🚪', `Ending the consumer loop`, {
-                isRunning: isRunning(),
-                isStale: isStale(),
-                msFromBatchStart: new Date().valueOf() - batchStartTimer.valueOf(),
-            })
-            await heartbeat()
-            return
-        }
     } finally {
         queue.pluginsServer.statsd?.timing(`kafka_queue.${loggingKey}`, batchStartTimer)
         transaction.finish()
